pyproject.toml
[project]
name = "jarvis-mcp"
version = "0.1.0"
description = "MCP proxy that aggregates multiple MCP servers behind 2 synthetic tools"
requires-python = ">=3.11"
dependencies = [
    "fastmcp[code-mode]>=3.2.0",
    "mcp>=1.0",
    "py-key-value-aio[disk]",
    "textual>=0.83",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "coverage[toml]>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jarvis"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
